Potion
The free Alpha version of Potion that checks the weather for the next 12 hours, the stocks for yesterday (or Friday depending on the day), and also returns the top 50 current trending Twitter Trends. Example code made using flask. Don't forget to add your API keys for openweathermap and Alphavantage before use.